exercise 2 - Game Context
