+++ src/components/App.js
@@ -1,25 +1,21 @@
-import React from "react";
+import React, { useContext } from "react";
 import { BrowserRouter as Router, Route } from "react-router-dom";
 
 import GlobalStyles from "./GlobalStyles";
 import Home from "./Home";
 import Game from "./Game";
 import usePersistedState from "../hooks/usePersistedState";
-import useInterval from "../hooks/use-interval.hook";
-
-export const items = [
-  { id: "cursor", name: "Cursor", cost: 10, value: 1 },
-  { id: "grandma", name: "Grandma", cost: 100, value: 10 },
-  { id: "farm", name: "Farm", cost: 1000, value: 80 },
-];
+import { GameContext } from "./GameContext";
 
 function App() {
-  const [numCookies, setNumCookies] = React.useState(1000);
-  const [purchasedItems, setPurchasedItems] = React.useState({
-    cursor: 0,
-    grandma: 0,
-    farm: 0,
-  });
+  const {
+    setNumCookies,
+    setPurchasedItems,
+    numCookies,
+    purchasedItems,
+    calculateCookiesPerSecond,
+  } = useContext(GameContext);
+
   // const [numCookies, setNumCookies] = React.usePersistedState(
   //   "numCookies",
   //   1000
@@ -29,22 +25,6 @@ function App() {
   //   grandma: 0,
   //   farm: 0,
   // });
-  const calculateCookiesPerSecond = (purchasedItems) => {
-    return Object.keys(purchasedItems).reduce((acc, itemId) => {
-      const numOwned = purchasedItems[itemId];
-      const item = items.find((item) => item.id === itemId);
-      const value = item.value;
-
-      return acc + value * numOwned;
-    }, 0);
-  };
-  // let numOfGeneratedCookies = 0
-  useInterval(() => {
-    const numOfGeneratedCookies = calculateCookiesPerSecond(purchasedItems);
-
-    setNumCookies(numCookies + numOfGeneratedCookies);
-    window.localStorage.setItem("num-cookies", numCookies);
-  }, 1000);
 
   return (
     <>
