+++ src/index.js
@@ -2,7 +2,13 @@ import React from "react";
 import ReactDOM from "react-dom";
 
 import App from "./components/App";
+import { GameProvider } from "./components/GameContext";
 
 const rootElement = document.getElementById("root");
 
-ReactDOM.render(<App />, rootElement);
+ReactDOM.render(
+  <GameProvider>
+    <App />
+  </GameProvider>,
+  rootElement
+);
